requirements.txt
boto3==1.5.19
python-consul==0.7.2
